MAINT: Fix bad merge
